Refactor PokemonDetail
