+++ frontend/components/pokemonDetail/Evolutions.module.scss
@@ -1,0 +1,34 @@
+@use '@carbon/styles/scss/theme';
+@use '@carbon/layout';
+
+.evolutionsTitle {
+  padding-top: layout.$spacing-06;
+  color: theme.$text-primary;
+}
+
+.grid {
+  display: grid;
+  width: 100%;
+  padding: layout.$spacing-04 0;
+  grid-template-columns: repeat(4, 25%);
+  .gridGap {
+    &:first-child {
+      padding-left: 0;
+    }
+    &:last-child {
+      padding-right: 0;
+    }
+    padding: 0 layout.$spacing-03;
+  }
+}
+
+@media (max-width: 768px) {
+  .grid {
+    grid-template-columns: repeat(3, 33.33%);
+  }
+}
+@media (max-width: 480px) {
+  .grid {
+    grid-template-columns: repeat(2, 50%);
+  }
+}
